+++ tests/test_imports.py
@@ -1,0 +1,12 @@
+"""
+Does `admonitions` import without fail?
+"""
+
+
+def test_imports():
+    """
+    Simply import `admonitions`, and return true if no errors are thrown.
+    """
+    import admonitions
+
+    return True