+++ CHANGELOG.md
@@ -2,6 +2,5 @@
 
 ## Version 0.1 (development)
 
-- Feature A added
-- FIX: nasty bug #1729 fixed
-- add your changes here!
+- Initial release for rich-admonitions!
+- Provides one type, `Admonition`, which implements `rich`'s console protocol